Create drupal-cgi-worker.mjs
